backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-Talisman==1.1.0
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
PyMySQL==1.1.0
pytest==8.2.1
pytest-cov==5.0.0
alembic==1.13.1
marshmallow==3.21.1
